requirements.txt
python-dotenv==1.0.1
requests==2.31.0
urllib3==1.26.8
python-telegram-bot==13.0




